+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.1.2
 Release TBD
 - Add config options to hide associate tag in links based on HTTP header and admin status
+- Add Prime logo to detail page if applicable
 
 ## Version 0.1.1
 Release 2017-10-16